Move IME simulation out of macOS and revert CJK heuristics
